+++ UpgradeWorld/actions/vegetation/RemoveVegetation.cs
@@ -36,7 +36,7 @@ public class RemoveVegetation : ZoneOperation
     if (zdos == null) return;
     foreach (var zdo in zdos)
     {
-      if (!Args.Roll()) continue;
+      if (!Args.RollAmount()) continue;
       if (!Hashes.Contains(zdo.m_prefab)) continue;
       AddPin(zdo.GetPosition());
       Helper.RemoveZDO(zdo);