+++ UpgradeWorld/actions/vegetation/VegetationOperation.cs
@@ -14,11 +14,13 @@ public abstract class VegetationOperation : ZoneOperation
     return [.. AllIds().Where(id => (ids.Count == 0 || ids.Contains(id)) && (ignore.Count == 0 || !ignore.Contains(id)))];
   }
   public HashSet<string> Ids = [];
+  private float Amount = 1f;
   private static List<ZoneSystem.ZoneVegetation> OriginalVegetation = [];
   private static List<ZoneSystem.ZoneVegetation> Vegetation = [];
   public VegetationOperation(Terminal context, HashSet<string> ids, FiltererParameters args) : base(context, args)
   {
     args.TargetZones = TargetZones.Generated;
+    Amount = args.Amount;
     Ids = ids;
   }
   protected override void OnStart()
@@ -28,6 +30,9 @@ public abstract class VegetationOperation : ZoneOperation
     Vegetation = [.. ZoneSystem.instance.m_vegetation.Select(veg => veg.Clone())];
     if (Ids.Count > 0)
       Set(GetWithOnlyIds(Ids, true));
+    if (Amount != 1f)
+      SetAmount(Amount);
+
   }
   protected void OverrideVegetation()
   {
@@ -37,13 +42,19 @@ public abstract class VegetationOperation : ZoneOperation
   {
     ZoneSystem.instance.m_vegetation = OriginalVegetation;
   }
-  public static bool[] GetWithIds(HashSet<string> ids, bool value) =>
-    [.. Vegetation.Select(veg => ids.Contains(veg.m_prefab.name) ? value : veg.m_enable)];
-  public static bool[] GetWithOnlyIds(HashSet<string> ids, bool value) =>
+  private static bool[] GetWithOnlyIds(HashSet<string> ids, bool value) =>
     [.. Vegetation.Select(veg => ids.Contains(veg.m_prefab.name) ? value : !value)];
-  public static void Set(bool[] values)
+  private static void Set(bool[] values)
   {
     for (var i = 0; i < Vegetation.Count; i++) Vegetation[i].m_enable = values[i];
+  }
+  private static void SetAmount(float amount)
+  {
+    foreach (var veg in Vegetation)
+    {
+      veg.m_min += amount;
+      veg.m_max += amount;
+    }
   }
 
   public static void Register(string name)